Export list_save() implementation
