+++ src/param/list/param_list_slash.c
@@ -6,7 +6,6 @@
  */
 
 #include <stdio.h>
-#include <stdlib.h>
 #include <string.h>
 #include <inttypes.h>
 #include <slash/slash.h>
@@ -218,18 +217,7 @@ static int list_add(struct slash *slash)
 slash_command_sub(list, add, list_add, "<name> <id> <type>", NULL);
 
 
-int param_sort_fnc(const void* p1, const void* p2) {
-
-    param_t* param1 = *(param_t**)p1;
-    param_t* param2 = *(param_t**)p2;
-
-    if (param1->id > param2->id) return 1;
-    if (param1->id < param2->id) return -1;
-    return 0;
-}
-
-
-static int list_save(struct slash *slash) {
+static int list_save_cmd(struct slash *slash) {
 
     char * filename = NULL;
     int node = slash_dfl_node;
@@ -247,74 +235,9 @@ static int list_save(struct slash *slash) {
        return SLASH_EINVAL;
     }
 
-    FILE * out = stdout;
-
-    if (filename) {
-       FILE * fd = fopen(filename, "w");
-        if (fd) {
-            out = fd;
-            printf("Writing to file %s\n", filename);
-        }
-    }
-
-    param_t * param;
-    param_list_iterator i = {};
-    param_t* param_sorted[1024];
-    int param_cnt = 0;
-
-    while ((param = param_list_iterate(&i)) != NULL) {
-
-        if ((node >= 0) && (param->node != node)) {
-            continue;
-        }
-        param_sorted[param_cnt] = param;
-        param_cnt++;
-    };
-
-    qsort(param_sorted, param_cnt, sizeof(param_sorted[0]), param_sort_fnc);
-
-    for (int i = 0; i < param_cnt; i++) {
-        fprintf(out, "list add ");
-        if (param_sorted[i]->array_size > 1) {
-            fprintf(out, "-a %u ", param_sorted[i]->array_size);
-        }
-        if ((param_sorted[i]->docstr != NULL) && (strlen(param_sorted[i]->docstr) > 0)) {
-            fprintf(out, "-c \"%s\" ", param_sorted[i]->docstr);
-        }
-        if ((param_sorted[i]->unit != NULL) && (strlen(param_sorted[i]->unit) > 0)) {
-            fprintf(out, "-u \"%s\" ", param_sorted[i]->unit);
-        }
-        if (param_sorted[i]->node != 0 && !skip_node) {
-            fprintf(out, "-n %u ", param_sorted[i]->node);
-        }
-        
-        if (param_sorted[i]->mask > 0) {
-            unsigned int mask = param_sorted[i]->mask;
-        
-            list_add_output(mask, out);
-            list_add_output_user_flags(mask,out);
-        }
-
-        if (param_sorted[i]->vmem != NULL && param_sorted[i]->vmem->type > 0) {
-            fprintf(out, "-v %u ", param_sorted[i]->vmem->type);
-        }
-
-        fprintf(out, "%s %u ", param_sorted[i]->name, param_sorted[i]->id);
-
-        char typestr[10];
-        param_type_str(param_sorted[i]->type, typestr, 10);
-        fprintf(out, "%s\n", typestr);
-
-    }
-
-    if (out != stdout) {
-        fflush(out);
-        fclose(out);
-    }
-
+    param_list_save(filename, node, skip_node);
 
     optparse_del(parser);
     return SLASH_SUCCESS;
 }
-slash_command_sub(list, save, list_save, "", "Save parameters");
-
+slash_command_sub(list, save, list_save_cmd, "", "Save parameters");
